Split the module in 2: - one for the business feature (account_multicurrency_revaluation) - one for the report feature (account_multicurrency_revaluation_report)
I move the report feature to a dedicated module because else the whole module could not be upgraded because of this pull request: https://github.com/OCA/webkit-tools/pull/10
Indeed, the module previously depended on the module base_headers_webkit. Now, only the module for the report (account_multicurrency_revaluation_report) will depend on it.

The reference to the currency type has been removed because it has been removed from the core addons in v8.

The code is not fully ported to the new API yet.