account_invoice_accrual: simplify and partially migrate to new api
Main motivation for this change is that the override of account_reversal
is fragile, error prone and may not work in all situations (in particular
when reversal are done with another mean than the wizard). So it's
better to leave the link to accrual_move_id even after reversal: it
does no harm.